Refactors for optimisation, readability and correctness
Makes sure that the sampling uses the actual maximum value encountered, not the maximum from the previous sample.
Calculates, rather than searches for the bucket to increment in the histogram map.
Commands moved into separate packages, as per golang advice.
Test written for some aspects of the scale / reverse scale functions in spector.